Fix compilation error
Fixes:
westeros-nested.cpp:1104:46: error: cast from ‘void*’ to ‘uint32_t’ {aka ‘unsigned int’} loses precision [-fpermissive]
 1104 |                                    (uint32_t)deviceBuffer,
      |                                              ^~~~~~~~~~~~

Note: The API might not work on 64-bit platforms, it needs to be double checked.

+++ westeros-nested.cpp
@@ -1101,7 +1101,7 @@ void WstNestedConnectionAttachAndCommitDevice( WstNestedConnection *nc,
       struct wl_buffer *buffer;
       
       buffer= wl_sb_create_buffer( nc->sb, 
-                                   (uint32_t)deviceBuffer, 
+                                   (uintptr_t)deviceBuffer,
                                    width, 
                                    height, 
                                    stride,
